Revert "gcd should work for decimals too"
This reverts commit abd9090509c87fd5e5d090dfc291f7b67e0a39e5.

+++ Voxel/Common.h
@@ -127,7 +127,7 @@ Ptr<T> byteAlloc(unsigned long sizeInBytes)
   return Ptr<T>((T *)new uint8_t[sizeInBytes], [](T *d) { delete[]((uint8_t *)d);  });
 }
 
-float VOXEL_EXPORT gcd(float n, float m);
+uint VOXEL_EXPORT gcd(uint n, uint m);
 
 // This returns nearest 'n = 2^m' such that 2^(m - 1) < value < 2^m
 unsigned int VOXEL_EXPORT nearestPowerOf2(unsigned int value, unsigned int &index);
